text_summary/requirement.txt
conda install pytorch torchvision torchaudio -c pytorch-nightly
tqdm
numpy
pandas
matplotlib
sentencepiece
pytorch ignite